+++ src/Plugin/SensiolabsInsight.php
@@ -62,29 +62,13 @@ class SensiolabsInsight extends Plugin
 
         if (\array_key_exists('auth_token', $options)) {
             $this->authToken = $options['auth_token'];
-            /** @deprecated Option "api_token" is deprecated and will be deleted in version 2.0. Use the option "auth_token" instead. */
-        } elseif (\array_key_exists('api_token', $options)) {
-            $builder->logWarning(
-                '[DEPRECATED] Option "api_token" is deprecated and will be deleted in version 2.0. Use the option "auth_token" instead.'
-            );
-
-            $this->authToken = $options['api_token'];
         }
 
         if (array_key_exists('project_uuid', $options)) {
             $this->projectUuid = $options['project_uuid'];
         }
 
-        /** @deprecated Option "executable" is deprecated and will be deleted in version 2.0. Use the option "binary_path" and "binary_name" instead. */
-        if (array_key_exists('executable', $options)) {
-            $this->builder->logWarning(
-                '[DEPRECATED] Option "executable" is deprecated and will be deleted in version 2.0. Use the option "binary_path" and "binary_name" instead.'
-            );
-
-            $this->executable = $this->builder->interpolate($options['executable']);
-        } else {
-            $this->executable = $this->findBinary('insight');
-        }
+        $this->executable = $this->findBinary('insight');
     }
 
     /**